UIE-165 error.ts more decoupling for Environments code sharing - shifted most NotificationsProvider bits into notifications-provider.ts module. - ts error fix and minor cleanup.
