+++ packages/notifications/src/index.ts
@@ -1,2 +1,3 @@
 export * from './error';
+export * from './notifications-provider';
 export * from './useNotifications';
