+++ packages/notifications/src/useNotifications.test.tsx
@@ -2,12 +2,8 @@ import { ErrorBoundary } from '@terra-ui-packages/components';
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import {
-  NotificationsContextProvider,
-  NotificationsProvider,
-  text,
-  useNotificationsFromContext,
-} from './useNotifications';
+import { NotificationsProvider } from './notifications-provider';
+import { NotificationsContextProvider, text, useNotificationsFromContext } from './useNotifications';
 
 describe('useNotificationsFromContext', () => {
   it('gets notifications provider from context', () => {
